+++ code/addons/a11y/src/manager.test.tsx
@@ -10,6 +10,7 @@ import './manager';
 
 vi.mock('storybook/manager-api');
 const mockedApi = vi.mocked<api.API>(api as any);
+mockedApi.useStorybookApi = vi.fn(() => ({ getSelectedPanel: vi.fn() }));
 mockedApi.useAddonState = vi.fn();
 const mockedAddons = vi.mocked(api.addons);
 const registrationImpl = mockedAddons.register.mock.calls[0][1];
@@ -44,21 +45,18 @@ describe('A11yManager', () => {
 
     // when / then
     expect(title()).toMatchInlineSnapshot(`
-      <div>
-        <Spaced
-          col={1}
-        >
-          <span
-            style={
-              {
-                "display": "inline-block",
-                "verticalAlign": "middle",
-              }
-            }
-          >
-            Accessibility
-          </span>
-        </Spaced>
+      <div
+        style={
+          {
+            "alignItems": "center",
+            "display": "flex",
+            "gap": 6,
+          }
+        }
+      >
+        <span>
+          Accessibility
+        </span>
       </div>
     `);
   });
@@ -77,26 +75,24 @@ describe('A11yManager', () => {
 
     // when / then
     expect(title()).toMatchInlineSnapshot(`
-      <div>
-        <Spaced
-          col={1}
+      <div
+        style={
+          {
+            "alignItems": "center",
+            "display": "flex",
+            "gap": 6,
+          }
+        }
+      >
+        <span>
+          Accessibility
+        </span>
+        <Badge
+          compact={true}
+          status="neutral"
         >
-          <span
-            style={
-              {
-                "display": "inline-block",
-                "verticalAlign": "middle",
-              }
-            }
-          >
-            Accessibility
-          </span>
-          <Badge
-            status="neutral"
-          >
-            3
-          </Badge>
-        </Spaced>
+          3
+        </Badge>
       </div>
     `);
   });