+++ code/core/src/builder-manager/index.ts
@@ -9,6 +9,7 @@ import { globalExternals } from '@fal-works/esbuild-plugin-global-externals';
 import { pnpPlugin } from '@yarnpkg/esbuild-plugin-pnp';
 import sirv from 'sirv';
 
+import { BROWSER_TARGETS } from '../shared/constants/environments-support';
 import type {
   BuilderBuildResult,
   BuilderFunction,
@@ -66,10 +67,17 @@ export const getConfig: ManagerBuilder['getConfig'] = async (options) => {
       '.eot': 'dataurl',
       '.ttf': 'dataurl',
     },
-    target: ['chrome100', 'safari15', 'firefox91'],
+    target: BROWSER_TARGETS,
     platform: 'browser',
     bundle: true,
-    minify: true,
+    minify: false,
+    minifyWhitespace: false,
+    minifyIdentifiers: false,
+    minifySyntax: false,
+    metafile: true,
+
+    // treeShaking: true,
+
     sourcemap: false,
     conditions: ['browser', 'module', 'default'],
 
@@ -170,6 +178,13 @@ const starter: StarterFunction = async function* starterGeneratorFn({
   );
 
   const { cssFiles, jsFiles } = await readOrderedFiles(addonsDir, compilation?.outputFiles);
+
+  if (compilation.metafile && options.outputDir) {
+    await writeFile(
+      join(options.outputDir, 'metafile.json'),
+      JSON.stringify(compilation.metafile, null, 2)
+    );
+  }
 
   // Build additional global values
   const globals: Record<string, any> = await buildFrameworkGlobalsFromOptions(options);