+++ code/core/src/core-server/stores/status.ts
@@ -3,21 +3,18 @@ import { UNIVERSAL_STATUS_STORE_OPTIONS } from '../../shared/status-store';
 import { UniversalStore } from '../../shared/universal-store';
 
 const statusStore = createStatusStore({
-  universalStatusStore:
+  universalStatusStore: UniversalStore.create({
+    ...UNIVERSAL_STATUS_STORE_OPTIONS,
     /*
-    This is a temporary workaround, to ensure that the store is not created in the
-    vitest sub-process in addon-test, even though it imports from core-server
-    If it was created in the sub-process, it would try to connect to the leader in the dev server
-    before it was ready.
-    This will be fixed when we do the planned UniversalStore v0.2.
+      This is a temporary workaround, to ensure that the store is not created in the
+      vitest sub-process in addon-vitest, even though it imports from core-server
+      If it was created in the sub-process, it would try to connect to the leader in the dev server
+      before it was ready.
+      This will be fixed when we do the planned UniversalStore v0.2.
     */
-    process.env.VITEST !== 'true'
-      ? UniversalStore.create({
-          ...UNIVERSAL_STATUS_STORE_OPTIONS,
-          leader: true,
-        })
-      : ({} as any),
+    leader: process.env.VITEST_CHILD_PROCESS !== 'true',
+  }),
   environment: 'server',
 });
 
-export const { fullStatusStore, getStatusStoreByTypeId } = statusStore;
+export const { fullStatusStore, getStatusStoreByTypeId, universalStatusStore } = statusStore;