+++ code/core/src/core-server/stores/test-provider.ts
@@ -3,20 +3,18 @@ import { UNIVERSAL_TEST_PROVIDER_STORE_OPTIONS } from '../../shared/test-provide
 import { UniversalStore } from '../../shared/universal-store';
 
 const testProviderStore = createTestProviderStore({
-  universalTestProviderStore:
+  universalTestProviderStore: UniversalStore.create({
+    ...UNIVERSAL_TEST_PROVIDER_STORE_OPTIONS,
     /*
-    This is a temporary workaround, to ensure that the store is not created in the
-    vitest sub-process in addon-test, even though it imports from core-server
-    If it was created in the sub-process, it would try to connect to the leader in the dev server
-    before it was ready.
-    This will be fixed when we do the planned UniversalStore v0.2.
-    */
-    process.env.VITEST !== 'true'
-      ? UniversalStore.create({
-          ...UNIVERSAL_TEST_PROVIDER_STORE_OPTIONS,
-          leader: true,
-        })
-      : ({} as any),
+            This is a temporary workaround, to ensure that the store is not created in the
+            vitest sub-process in addon-vitest, even though it imports from core-server
+            If it was created in the sub-process, it would try to connect to the leader in the dev server
+            before it was ready.
+            This will be fixed when we do the planned UniversalStore v0.2.
+          */
+    leader: process.env.VITEST_CHILD_PROCESS !== 'true',
+  }),
 });
 
-export const { fullTestProviderStore, getTestProviderStoreById } = testProviderStore;
+export const { fullTestProviderStore, getTestProviderStoreById, universalTestProviderStore } =
+  testProviderStore;