+++ code/core/src/core-server/utils/StoryIndexGenerator.ts
@@ -25,11 +25,12 @@ import type {
 import { findUp } from 'find-up';
 import picocolors from 'picocolors';
 import slash from 'slash';
-import { sortStoriesV7, userOrAutoTitleFromSpecifier } from 'storybook/preview-api';
 import invariant from 'tiny-invariant';
 import { dedent } from 'ts-dedent';
 import * as TsconfigPaths from 'tsconfig-paths';
 
+import { userOrAutoTitleFromSpecifier } from '../../preview-api/modules/store/autoTitle';
+import { sortStoriesV7 } from '../../preview-api/modules/store/sortStories';
 import { IndexingError, MultipleIndexingError } from './IndexingError';
 import { autoName } from './autoName';
 import { type IndexStatsSummary, addStats } from './summarizeStats';
@@ -400,7 +401,7 @@ export class StoryIndexGenerator {
     //  a) autodocs is globally enabled
     //  b) we have autodocs enabled for this file
     const hasAutodocsTag = entries.some((entry) => entry.tags.includes(AUTODOCS_TAG));
-    const createDocEntry = hasAutodocsTag && !!this.options.docs.autodocs;
+    const createDocEntry = hasAutodocsTag;
 
     if (createDocEntry && this.options.build?.test?.disableAutoDocs !== true) {
       const name = this.options.docs.defaultName ?? 'Docs';
@@ -592,10 +593,7 @@ export class StoryIndexGenerator {
       }
 
       // If you link a file to a tagged CSF file, you have probably made a mistake
-      if (
-        worseEntry.tags?.includes(AUTODOCS_TAG) &&
-        !(this.options.docs.autodocs === true || projectTags?.includes(AUTODOCS_TAG))
-      ) {
+      if (worseEntry.tags?.includes(AUTODOCS_TAG) && !projectTags?.includes(AUTODOCS_TAG)) {
         throw new IndexingError(
           `You created a component docs page for '${worseEntry.title}', but also tagged the CSF file with '${AUTODOCS_TAG}'. This is probably a mistake.`,
           [betterEntry.importPath, worseEntry.importPath]
@@ -768,7 +766,6 @@ export class StoryIndexGenerator {
   getProjectTags(previewCode?: string) {
     let projectTags = [] as Tag[];
     const defaultTags = ['dev', 'test'];
-    const extraTags = this.options.docs.autodocs === true ? [AUTODOCS_TAG] : [];
     if (previewCode) {
       try {
         const projectAnnotations = loadConfig(previewCode).parse();
@@ -789,7 +786,7 @@ export class StoryIndexGenerator {
         `);
       }
     }
-    return [...defaultTags, ...projectTags, ...extraTags];
+    return [...defaultTags, ...projectTags];
   }
 
   // Get the story file names in "imported order"