+++ code/core/src/instrumenter/instrumenter.ts
@@ -11,21 +11,12 @@ import type { StoryId } from 'storybook/internal/types';
 import { global } from '@storybook/global';
 
 import { processError } from '@vitest/utils/error';
-import { addons } from 'storybook/preview-api';
 
+import { EVENTS } from './EVENTS';
+import { addons } from './preview-api';
 import type { Call, CallRef, ControlStates, LogItem, Options, State, SyncPayload } from './types';
 import { CallStates } from './types';
 import './typings.d.ts';
-
-export const EVENTS = {
-  CALL: 'storybook/instrumenter/call',
-  SYNC: 'storybook/instrumenter/sync',
-  START: 'storybook/instrumenter/start',
-  BACK: 'storybook/instrumenter/back',
-  GOTO: 'storybook/instrumenter/goto',
-  NEXT: 'storybook/instrumenter/next',
-  END: 'storybook/instrumenter/end',
-};
 
 type PatchedObj<TObj extends Record<string, unknown>> = {
   [Property in keyof TObj]: TObj[Property] & { __originalFn__: TObj[Property] };
@@ -245,30 +236,32 @@ export class Instrumenter {
     };
 
     // Support portable stories where addons are not available
-    (addons ? addons.ready() : Promise.resolve()).then(() => {
-      this.channel = addons.getChannel();
+    if (addons) {
+      addons.ready().then(() => {
+        this.channel = addons.getChannel();
 
-      // A forceRemount might be triggered for debugging (on `start`), or elsewhere in Storybook.
-      this.channel.on(FORCE_REMOUNT, resetState);
+        // A forceRemount might be triggered for debugging (on `start`), or elsewhere in Storybook.
+        this.channel.on(FORCE_REMOUNT, resetState);
 
-      // Start with a clean slate before playing after a remount, and stop debugging when done.
-      this.channel.on(STORY_RENDER_PHASE_CHANGED, renderPhaseChanged);
+        // Start with a clean slate before playing after a remount, and stop debugging when done.
+        this.channel.on(STORY_RENDER_PHASE_CHANGED, renderPhaseChanged);
 
-      // Trash non-retained state and clear the log when switching stories, but not on initial boot.
-      this.channel.on(SET_CURRENT_STORY, () => {
-        if (this.initialized) {
-          this.cleanup();
-        } else {
-          this.initialized = true;
-        }
+        // Trash non-retained state and clear the log when switching stories, but not on initial boot.
+        this.channel.on(SET_CURRENT_STORY, () => {
+          if (this.initialized) {
+            this.cleanup();
+          } else {
+            this.initialized = true;
+          }
+        });
+
+        this.channel.on(EVENTS.START, start(this.channel));
+        this.channel.on(EVENTS.BACK, back(this.channel));
+        this.channel.on(EVENTS.GOTO, goto(this.channel));
+        this.channel.on(EVENTS.NEXT, next(this.channel));
+        this.channel.on(EVENTS.END, end);
       });
-
-      this.channel.on(EVENTS.START, start(this.channel));
-      this.channel.on(EVENTS.BACK, back(this.channel));
-      this.channel.on(EVENTS.GOTO, goto(this.channel));
-      this.channel.on(EVENTS.NEXT, next(this.channel));
-      this.channel.on(EVENTS.END, end);
-    });
+    }
   }
 
   getState(storyId: StoryId) {
@@ -355,12 +348,14 @@ export class Instrumenter {
       (acc, key) => {
         const descriptor = getPropertyDescriptor(obj, key);
         if (typeof descriptor?.get === 'function') {
-          const getter = () => descriptor?.get?.bind(obj)?.();
-          Object.defineProperty(acc, key, {
-            get: () => {
-              return this.instrument(getter(), { ...options, path: path.concat(key) }, depth);
-            },
-          });
+          if (descriptor.configurable) {
+            const getter = () => descriptor?.get?.bind(obj)?.();
+            Object.defineProperty(acc, key, {
+              get: () => {
+                return this.instrument(getter(), { ...options, path: path.concat(key) }, depth);
+              },
+            });
+          }
           return acc;
         }
 
@@ -456,7 +451,7 @@ export class Instrumenter {
 
     // TODO This function should not needed anymore, as the channel already serializes values with telejson
     // Possibly we need to add HTMLElement support to telejson though
-    // Keeping this function here, as removing it means we need to refactor the deserializing that happens in addon-interactions
+    // Keeping this function here, as removing it means we need to refactor the deserializing that happens in core interactions
     const maximumDepth = 25; // mimicks the max depth of telejson
     const serializeValues = (value: any, depth: number, seen: unknown[]): any => {
       if (seen.includes(value)) {
@@ -576,8 +571,8 @@ export class Instrumenter {
       const finalArgs = actualArgs.map((arg: any) => {
         // We only want to wrap plain functions, not objects.
 
-        // We only want to wrap plain functions, not objects.
-        if (typeof arg !== 'function' || Object.keys(arg).length) {
+        // We only want to wrap plain functions, not objects or classes.
+        if (typeof arg !== 'function' || isClass(arg) || Object.keys(arg).length) {
           return arg;
         }
 
@@ -743,4 +738,53 @@ function getPropertyDescriptor<T>(obj: T, propName: keyof T) {
     target = Object.getPrototypeOf(target);
   }
   return undefined;
+}
+
+export function isClass(obj: unknown) {
+  // if not a function, return false.
+
+  // if not a function, return false.
+  if (typeof obj !== 'function') {
+    return false;
+  }
+
+  // ⭐ is a function, has a prototype, and can't be deleted!
+
+  // ⭐ although a function's prototype is writable (can be reassigned),
+  //   it's not configurable (can't update property flags), so it
+  //   will remain writable.
+  //
+  // ⭐ a class's prototype is non-writable.
+  //
+  // Table: property flags of function/class prototype
+  // ---------------------------------
+  //   prototype  write  enum  config
+  // ---------------------------------
+  //   function     v      .      .
+  //   class        .      .      .
+  // ---------------------------------
+
+  // ⭐ is a function, has a prototype, and can't be deleted!
+
+  // ⭐ although a function's prototype is writable (can be reassigned),
+  //   it's not configurable (can't update property flags), so it
+  //   will remain writable.
+  //
+  // ⭐ a class's prototype is non-writable.
+  //
+  // Table: property flags of function/class prototype
+  // ---------------------------------
+  //   prototype  write  enum  config
+  // ---------------------------------
+  //   function     v      .      .
+  //   class        .      .      .
+  // ---------------------------------
+  const descriptor = Object.getOwnPropertyDescriptor(obj, 'prototype');
+
+  // every method shorthand version has no prototype
+  if (!descriptor) {
+    return false;
+  }
+
+  return !descriptor.writable;
 }