+++ code/core/src/preview-api/index.ts
@@ -76,6 +76,13 @@ export type { PropDescriptor } from './store';
 
 /** STORIES API */
 export { StoryStore, type Report, ReporterAPI } from './store';
-export { Preview, PreviewWeb, PreviewWithSelection, UrlStore, WebView } from './preview-web';
+export {
+  Preview,
+  PreviewWeb,
+  PreviewWithSelection,
+  UrlStore,
+  WebView,
+  emitTransformCode,
+} from './preview-web';
 export type { SelectionStore, View } from './preview-web';
-export { getComposedCoreAnnotations, getCoreAnnotations } from './core-annotations';
+export { getCoreAnnotations } from './core-annotations';