+++ code/core/src/shared/preview/csf4.ts
@@ -1,3 +1,4 @@
+import componentTestingAnnotations from 'storybook/internal/component-testing/preview';
 import type {
   ComponentAnnotations,
   Meta,
@@ -9,8 +10,13 @@ import type {
 import type { NormalizedProjectAnnotations } from 'storybook/internal/types';
 
 import actionAnnotations from 'storybook/actions/preview';
+import backgroundsAnnotations from 'storybook/backgrounds/preview';
+import highlightAnnotations from 'storybook/highlight/preview';
+import measureAnnotations from 'storybook/measure/preview';
+import outlineAnnotations from 'storybook/outline/preview';
 import { composeConfigs, normalizeProjectAnnotations } from 'storybook/preview-api';
 import testAnnotations from 'storybook/test/preview';
+import viewportAnnotations from 'storybook/viewport/preview';
 
 /** Do not use, use the definePreview exported from the framework instead. */
 // eslint-disable-next-line @typescript-eslint/naming-convention, no-underscore-dangle
@@ -30,7 +36,20 @@ export function __definePreview<TRenderer extends Renderer>(
         // TODO: Remove coreAnnotations once csf-factories use prepareStory (as core annotations already come from it)
         composeConfigs([
           // @ts-expect-error CJS fallback
+          (viewportAnnotations.default ?? viewportAnnotations)(),
+          // @ts-expect-error CJS fallback
+          (backgroundsAnnotations.default ?? backgroundsAnnotations)(),
+          // @ts-expect-error CJS fallback
+          (highlightAnnotations.default ?? highlightAnnotations)(),
+
+          // @ts-expect-error CJS fallback
+          (measureAnnotations.default ?? measureAnnotations)(),
+          // @ts-expect-error CJS fallback
+          (outlineAnnotations.default ?? outlineAnnotations)(),
+          // @ts-expect-error CJS fallback
           (actionAnnotations.default ?? actionAnnotations)(),
+          // @ts-expect-error CJS fallback
+          (componentTestingAnnotations.default ?? componentTestingAnnotations)(),
           // @ts-expect-error CJS fallback
           (testAnnotations.default ?? testAnnotations)(),
           ...(addons ?? []),