+++ code/lib/cli-storybook/src/migrate.ts
@@ -8,7 +8,6 @@ import {
 import { listCodemods, runCodemod } from '@storybook/codemod';
 
 import { runFixes } from './automigrate';
-import { mdxToCSF } from './automigrate/fixes/mdx-to-csf';
 import { getStorybookData } from './automigrate/helpers/mainConfigFile';
 
 const logger = console;
@@ -31,50 +30,8 @@ export async function migrate(
   if (list) {
     listCodemods().forEach((key: any) => logger.log(key));
   } else if (migration) {
-    if (migration === 'mdx-to-csf' && !dryRun) {
-      const packageManager = JsPackageManagerFactory.getPackageManager();
-
-      const { configDir, mainConfig, mainConfigPath, storybookVersion, packageJson } =
-        await getStorybookData({
-          packageManager,
-          configDir: userSpecifiedConfigDir,
-        });
-
-      // GUARDS
-      if (!storybookVersion) {
-        throw new Error('Could not determine Storybook version');
-      }
-
-      if (!mainConfigPath) {
-        throw new Error('Could not determine main config path');
-      }
-
-      await runFixes({
-        fixes: [mdxToCSF],
-        configDir,
-        mainConfigPath,
-        packageManager,
-        mainConfig,
-        packageJson,
-        storybookVersion,
-        beforeVersion: storybookVersion,
-        isUpgrade: false,
-      });
-      await addStorybookBlocksPackage();
-    }
-
     await runCodemod(migration, { glob, dryRun, logger, rename, parser });
   } else {
     throw new Error('Migrate: please specify a migration name or --list');
   }
-}
-
-export async function addStorybookBlocksPackage() {
-  const packageManager = JsPackageManagerFactory.getPackageManager();
-  const packageJson = await packageManager.retrievePackageJson();
-  const versionToInstall = getStorybookVersionSpecifier(await packageManager.retrievePackageJson());
-  logger.info(`✅ Adding "@storybook/blocks" package`);
-  await packageManager.addDependencies({ installAsDevDependencies: true, packageJson }, [
-    `@storybook/blocks@${versionToInstall}`,
-  ]);
 }