+++ code/lib/create-storybook/src/generators/baseGenerator.ts
@@ -183,9 +183,6 @@ const getFrameworkDetails = (
 
 const stripVersions = (addons: string[]) => addons.map((addon) => getPackageDetails(addon)[0]);
 
-const hasInteractiveStories = (rendererId: SupportedRenderers) =>
-  ['react', 'angular', 'preact', 'svelte', 'vue3', 'html', 'solid', 'qwik'].includes(rendererId);
-
 const hasFrameworkTemplates = (framework?: SupportedFrameworks) => {
   if (!framework) {
     return false;
@@ -271,10 +268,6 @@ export async function baseGenerator(
 
   const compiler = webpackCompiler ? webpackCompiler({ builder }) : undefined;
 
-  const essentials = features.includes('docs')
-    ? '@storybook/addon-essentials'
-    : { name: '@storybook/addon-essentials', options: { docs: false } };
-
   const extraAddonsToInstall =
     typeof extraAddonPackages === 'function'
       ? await extraAddonPackages({
@@ -284,27 +277,26 @@ export async function baseGenerator(
       : extraAddonPackages;
 
   // TODO: change the semver range to '^4' when VTA 4 and SB 9 is released
-  extraAddonsToInstall.push('@chromatic-com/storybook@^4.0.0-0');
+  if (features.includes('test')) {
+    extraAddonsToInstall.push('@chromatic-com/storybook@^4.0.0-0');
+  }
+
+  // Add @storybook/addon-docs when docs feature is selected
+  if (features.includes('docs')) {
+    extraAddonsToInstall.push('@storybook/addon-docs');
+  }
 
   // added to main.js
   const addons = [
     ...(compiler ? [`@storybook/addon-webpack5-compiler-${compiler}`] : []),
-    essentials,
     ...stripVersions(extraAddonsToInstall),
   ].filter(Boolean);
 
   // added to package.json
   const addonPackages = [
-    '@storybook/addon-essentials',
-    '@storybook/blocks',
     ...(compiler ? [`@storybook/addon-webpack5-compiler-${compiler}`] : []),
     ...extraAddonsToInstall,
   ].filter(Boolean);
-
-  if (hasInteractiveStories(rendererId) && !features.includes('test')) {
-    addons.push('@storybook/addon-interactions');
-    addonPackages.push('@storybook/addon-interactions');
-  }
 
   const packageJson = await packageManager.retrievePackageJson();
   const installedDependencies = new Set(
@@ -431,7 +423,7 @@ export async function baseGenerator(
       frameworkPreviewParts,
       storybookConfigFolder: storybookConfigFolder as string,
       language,
-      rendererId,
+      frameworkPackage,
     });
   }