+++ code/lib/create-storybook/src/generators/configure.test.ts
@@ -90,7 +90,6 @@ describe('configureMain', () => {
       addons: [
         "%%path.dirname(require.resolve(path.join('@storybook/addon-essentials', 'package.json')))%%",
         "%%path.dirname(require.resolve(path.join('@storybook/preset-create-react-app', 'package.json')))%%",
-        "%%path.dirname(require.resolve(path.join('@storybook/addon-interactions', 'package.json')))%%",
       ],
       storybookConfigFolder: '.storybook',
       framework: {
@@ -114,8 +113,7 @@ describe('configureMain', () => {
         ],
         "addons": [
           path.dirname(require.resolve(path.join('@storybook/addon-essentials', 'package.json'))),
-          path.dirname(require.resolve(path.join('@storybook/preset-create-react-app', 'package.json'))),
-          path.dirname(require.resolve(path.join('@storybook/addon-interactions', 'package.json')))
+          path.dirname(require.resolve(path.join('@storybook/preset-create-react-app', 'package.json')))
         ],
         "framework": {
           "name": path.dirname(require.resolve(path.join('@storybook/react-webpack5', 'package.json')))
@@ -131,7 +129,7 @@ describe('configurePreview', () => {
     await configurePreview({
       language: SupportedLanguage.JAVASCRIPT,
       storybookConfigFolder: '.storybook',
-      rendererId: 'react',
+      frameworkPackage: '@storybook/react-vite',
     });
 
     const { calls } = vi.mocked(fsp.writeFile).mock;
@@ -139,7 +137,7 @@ describe('configurePreview', () => {
 
     expect(previewConfigPath).toEqual('./.storybook/preview.js');
     expect(previewConfigContent).toMatchInlineSnapshot(`
-      "/** @type { import('@storybook/react').Preview } */
+      "/** @type { import('@storybook/react-vite').Preview } */
       const preview = {
         parameters: {
           controls: {
@@ -159,7 +157,7 @@ describe('configurePreview', () => {
     await configurePreview({
       language: SupportedLanguage.TYPESCRIPT_4_9,
       storybookConfigFolder: '.storybook',
-      rendererId: 'react',
+      frameworkPackage: '@storybook/react-vite',
     });
 
     const { calls } = vi.mocked(fsp.writeFile).mock;
@@ -167,7 +165,7 @@ describe('configurePreview', () => {
 
     expect(previewConfigPath).toEqual('./.storybook/preview.ts');
     expect(previewConfigContent).toMatchInlineSnapshot(`
-      "import type { Preview } from '@storybook/react'
+      "import type { Preview } from '@storybook/react-vite'
 
       const preview: Preview = {
         parameters: {
@@ -189,7 +187,7 @@ describe('configurePreview', () => {
     await configurePreview({
       language: SupportedLanguage.TYPESCRIPT_4_9,
       storybookConfigFolder: '.storybook',
-      rendererId: 'react',
+      frameworkPackage: '@storybook/react-vite',
     });
     expect(fsp.writeFile).not.toHaveBeenCalled();
   });
@@ -198,7 +196,7 @@ describe('configurePreview', () => {
     await configurePreview({
       language: SupportedLanguage.TYPESCRIPT_4_9,
       storybookConfigFolder: '.storybook',
-      rendererId: 'angular',
+      frameworkPackage: '@storybook/angular',
       frameworkPreviewParts: {
         prefix: dedent`
         import { setCompodocJson } from "@storybook/addon-docs/angular";