+++ test-storybooks/ember-cli/.storybook/main.js
@@ -10,13 +10,8 @@ module.exports = {
   logLevel: 'debug',
   addons: [
     '@storybook/addon-a11y',
-    '@storybook/addon-storysource',
     '@storybook/addon-docs',
-    '@storybook/addon-controls',
-    '@storybook/addon-links',
-    '@storybook/addon-viewport',
-    '@storybook/addon-backgrounds',
-    '@storybook/addon-highlight',
+    '@storybook/addon-links'
   ],
   webpackFinal: async (config) => {
     config.module.rules.push({
@@ -38,7 +33,7 @@ module.exports = {
     return config;
   },
   core: {
-    channelOptions: { allowFunction: false, maxDepth: 10 },
+    channelOptions: { maxDepth: 10 },
     disableTelemetry: true,
   },
   staticDirs: ['../ember-output'],